requirements.txt
fastapi
uvicorn[standard]
requests
python-dotenv
validators
pyOpenSSL
python-whois
